Refactor openalea.core to not depends on vpltk and oalab.
