+++ core/conda/meta.yaml
@@ -1,4 +1,4 @@
-{% set version = "2.0.0" %}
+{% set version = "2.0.1" %}
 
 package:
   name: openalea.core
@@ -17,7 +17,7 @@ requirements:
     - openalea.deploy #==2.0.0
   run:
     - python
-    - openalea.vpltk
+    - ipykernel
 
 test:
   imports:
